Consolidating imaging pulling logic
