+++ pkg/kubelet/images/image_manager.go
@@ -27,25 +27,29 @@ import (
 	"k8s.io/kubernetes/pkg/util/flowcontrol"
 )
 
-// imagePuller pulls the image using Runtime.PullImage().
-// It will check the presence of the image, and report the 'image pulling',
-// 'image pulled' events correspondingly.
-type parallelImagePuller struct {
+// imageManager provides the functionalities for image pulling.
+type imageManager struct {
 	recorder record.EventRecorder
 	runtime  kubecontainer.Runtime
 	backOff  *flowcontrol.Backoff
+	// It will check the presence of the image, and report the 'image pulling', image pulled' events correspondingly.
+	puller imagePuller
 }
 
-// enforce compatibility.
-var _ imagePuller = &parallelImagePuller{}
+var _ ImageManager = &imageManager{}
 
-// NewImagePuller takes an event recorder and container runtime to create a
-// image puller that wraps the container runtime's PullImage interface.
-func newParallelImagePuller(recorder record.EventRecorder, runtime kubecontainer.Runtime, imageBackOff *flowcontrol.Backoff) imagePuller {
-	return &parallelImagePuller{
+func NewImageManager(recorder record.EventRecorder, runtime kubecontainer.Runtime, imageBackOff *flowcontrol.Backoff, serialized bool) ImageManager {
+	var puller imagePuller
+	if serialized {
+		puller = newSerialImagePuller(runtime)
+	} else {
+		puller = newParallelImagePuller(runtime)
+	}
+	return &imageManager{
 		recorder: recorder,
 		runtime:  runtime,
 		backOff:  imageBackOff,
+		puller:   puller,
 	}
 }
 
@@ -65,16 +69,16 @@ func shouldPullImage(container *api.Container, imagePresent bool) bool {
 }
 
 // records an event using ref, event msg.  log to glog using prefix, msg, logFn
-func (puller *parallelImagePuller) logIt(ref *api.ObjectReference, eventtype, event, prefix, msg string, logFn func(args ...interface{})) {
+func (m *imageManager) logIt(ref *api.ObjectReference, eventtype, event, prefix, msg string, logFn func(args ...interface{})) {
 	if ref != nil {
-		puller.recorder.Event(ref, eventtype, event, msg)
+		m.recorder.Event(ref, eventtype, event, msg)
 	} else {
 		logFn(fmt.Sprint(prefix, " ", msg))
 	}
 }
 
-// PullImage pulls the image for the specified pod and container.
-func (puller *parallelImagePuller) pullImage(pod *api.Pod, container *api.Container, pullSecrets []api.Secret) (error, string) {
+// EnsureImageExists pulls the image for the specified pod and container.
+func (m *imageManager) EnsureImageExists(pod *api.Pod, container *api.Container, pullSecrets []api.Secret) (error, string) {
 	logPrefix := fmt.Sprintf("%s/%s", pod.Name, container.Image)
 	ref, err := kubecontainer.GenerateContainerRef(pod, container)
 	if err != nil {
@@ -82,35 +86,37 @@ func (puller *parallelImagePuller) pullImage(pod *api.Pod, container *api.Contai
 	}
 
 	spec := kubecontainer.ImageSpec{Image: container.Image}
-	present, err := puller.runtime.IsImagePresent(spec)
+	present, err := m.runtime.IsImagePresent(spec)
 	if err != nil {
 		msg := fmt.Sprintf("Failed to inspect image %q: %v", container.Image, err)
-		puller.logIt(ref, api.EventTypeWarning, events.FailedToInspectImage, logPrefix, msg, glog.Warning)
+		m.logIt(ref, api.EventTypeWarning, events.FailedToInspectImage, logPrefix, msg, glog.Warning)
 		return ErrImageInspect, msg
 	}
 
 	if !shouldPullImage(container, present) {
 		if present {
 			msg := fmt.Sprintf("Container image %q already present on machine", container.Image)
-			puller.logIt(ref, api.EventTypeNormal, events.PulledImage, logPrefix, msg, glog.Info)
+			m.logIt(ref, api.EventTypeNormal, events.PulledImage, logPrefix, msg, glog.Info)
 			return nil, ""
 		} else {
 			msg := fmt.Sprintf("Container image %q is not present with pull policy of Never", container.Image)
-			puller.logIt(ref, api.EventTypeWarning, events.ErrImageNeverPullPolicy, logPrefix, msg, glog.Warning)
+			m.logIt(ref, api.EventTypeWarning, events.ErrImageNeverPullPolicy, logPrefix, msg, glog.Warning)
 			return ErrImageNeverPull, msg
 		}
 	}
 
 	backOffKey := fmt.Sprintf("%s_%s", pod.UID, container.Image)
-	if puller.backOff.IsInBackOffSinceUpdate(backOffKey, puller.backOff.Clock.Now()) {
+	if m.backOff.IsInBackOffSinceUpdate(backOffKey, m.backOff.Clock.Now()) {
 		msg := fmt.Sprintf("Back-off pulling image %q", container.Image)
-		puller.logIt(ref, api.EventTypeNormal, events.BackOffPullImage, logPrefix, msg, glog.Info)
+		m.logIt(ref, api.EventTypeNormal, events.BackOffPullImage, logPrefix, msg, glog.Info)
 		return ErrImagePullBackOff, msg
 	}
-	puller.logIt(ref, api.EventTypeNormal, events.PullingImage, logPrefix, fmt.Sprintf("pulling image %q", container.Image), glog.Info)
-	if err := puller.runtime.PullImage(spec, pullSecrets); err != nil {
-		puller.logIt(ref, api.EventTypeWarning, events.FailedToPullImage, logPrefix, fmt.Sprintf("Failed to pull image %q: %v", container.Image, err), glog.Warning)
-		puller.backOff.Next(backOffKey, puller.backOff.Clock.Now())
+	m.logIt(ref, api.EventTypeNormal, events.PullingImage, logPrefix, fmt.Sprintf("pulling image %q", container.Image), glog.Info)
+	errChan := make(chan error)
+	m.puller.pullImage(spec, pullSecrets, errChan)
+	if err := <-errChan; err != nil {
+		m.logIt(ref, api.EventTypeWarning, events.FailedToPullImage, logPrefix, fmt.Sprintf("Failed to pull image %q: %v", container.Image, err), glog.Warning)
+		m.backOff.Next(backOffKey, m.backOff.Clock.Now())
 		if err == RegistryUnavailable {
 			msg := fmt.Sprintf("image pull failed for %s because the registry is unavailable.", container.Image)
 			return err, msg
@@ -118,7 +124,7 @@ func (puller *parallelImagePuller) pullImage(pod *api.Pod, container *api.Contai
 			return ErrImagePull, err.Error()
 		}
 	}
-	puller.logIt(ref, api.EventTypeNormal, events.PulledImage, logPrefix, fmt.Sprintf("Successfully pulled image %q", container.Image), glog.Info)
-	puller.backOff.GC()
+	m.logIt(ref, api.EventTypeNormal, events.PulledImage, logPrefix, fmt.Sprintf("Successfully pulled image %q", container.Image), glog.Info)
+	m.backOff.GC()
 	return nil, ""
 }
